pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "discord-py>=2.6.4",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "pynacl>=1.6.0",
    "python-dotenv>=1.1.1",
    "spotipy>=2.25.1",
    "wavelink>=3.4.1",
    "yt-dlp>=2025.10.22",
]
